[⚙️ CHORE] 테이블 중복 삭제
