+++ src/main/resources/application-datasource.yml
@@ -13,7 +13,7 @@ spring:
     driver-class-name: org.h2.Driver
   jpa:
     hibernate:
-      ddl-auto: create-drop
+      ddl-auto: create
     show-sql: true
     database-platform: org.hibernate.dialect.H2Dialect
   h2:
